requirements.txt
# Core Dependencies
python-telegram-bot==20.7
aiohttp==3.9.3
playwright==1.40.0
fake-useragent==1.5.1
APScheduler==3.10.4
psutil==5.9.8
colorama==0.4.6
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.25
pydantic==2.5.2
schedule==1.2.0
pytz==2023.3

# Database
aiosqlite==0.19.0

# Utilities
tqdm==4.66.1
tabulate==0.9.0
pyfiglet==0.8.post1
rich==13.7.0

# Optional for Web
flask==3.0.0
flask-socketio==5.3.4

# Install Playwright browsers
# playwright install chromium